+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+```
+Added
+Changed
+Deprecated
+Removed
+Fixed
+Security
+```
+
 ### Changed
 
 - Refactor `enablePrefixer` to `addPrefixSelector`.