+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Removed 'obsidian' from plugin ID in `manifest.json`.
+- Removed 'obsidian' from plugin ID in `README.md` installation instructions.
 
 ### Added